Move SlugGenerator to Infrastructure
