+++ src/Catalog/Domain/Entity/Product/ValueObject/ProductSlug.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace App\Catalog\Domain\Entity\Product\ValueObject;
 
-use App\Shared\Domain\Util\SlugGenerator;
+use App\Shared\Infrastructure\String\SlugGenerator;
 use App\Shared\Domain\ValueObject\Slug;
 
 final class ProductSlug extends Slug
